Changed hierarchy of files
